feat: user metrics
