+++ pages/admin/me/metrics/dataservices.vue
@@ -1,0 +1,7 @@
+<template>
+  <MetricsAdminDataservicesPage :user="me" />
+</template>
+
+<script setup lang="ts">
+const me = useMe()
+</script>
