+++ pages/admin/me/metrics/reuses.vue
@@ -1,0 +1,9 @@
+<template>
+  <MetricsAdminReusesPage
+    :user="me"
+  />
+</template>
+
+<script setup lang="ts">
+const me = useMe()
+</script>
