+++ pages/admin/users/[uid]/metrics/dataservices.vue
@@ -1,0 +1,10 @@
+<template>
+  <MetricsAdminDataservicesPage
+    v-if="user"
+    :user
+  />
+</template>
+
+<script setup lang="ts">
+const { currentUser: user } = useCurrentOwned()
+</script>
