+++ pages/admin/users/[uid]/metrics/datasets.vue
@@ -1,0 +1,10 @@
+<template>
+  <MetricsAdminDatasetsPage
+    v-if="user"
+    :user
+  />
+</template>
+
+<script setup lang="ts">
+const { currentUser: user } = useCurrentOwned()
+</script>
