+++ pages/admin/users/[uid]/metrics/index.vue
@@ -1,0 +1,7 @@
+<template>
+  <MetricsAdminIndexPage :user="user" />
+</template>
+
+<script setup lang="ts">
+const { currentUser: user } = useCurrentOwned()
+</script>
