+++ pages/admin/users/[uid]/metrics/reuses.vue
@@ -1,0 +1,10 @@
+<template>
+  <MetricsAdminReusesPage
+    v-if="user"
+    :user
+  />
+</template>
+
+<script setup lang="ts">
+const { currentUser: user } = useCurrentOwned()
+</script>
